+++ tests/lib/test_nhentai.py
@@ -1,7 +1,7 @@
 import asyncio
 
-from comrade.lib.nhentai.parser import get_valid_nhentai_page, is_valid_page
-from comrade.lib.nhentai.structures import NHentaiGallery
+from comrade.lib.nhentai.parser import gallery_from_soup, is_valid_page
+from comrade.lib.nhentai.search import get_valid_nhentai_page
 
 
 def test_gallery_acquisition():
@@ -11,7 +11,7 @@ def test_gallery_acquisition():
 
     assert is_valid_page(page_soup)
 
-    gallery = NHentaiGallery.from_soup(page_soup)
+    gallery = gallery_from_soup(page_soup)
 
     assert gallery.gallery_id == gallery_id
     assert (
